Sonarqube fixes - Cast one of the operands of this multiplication operation to a ...
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1776894 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/poifs/filesystem/NDocumentInputStream.java
@@ -200,7 +200,7 @@ public final class NDocumentInputStream extends DocumentInputStream {
 		if (n < 0) {
 			return 0;
 		}
-		int new_offset = _current_offset + (int) n;
+		long new_offset = _current_offset + n;
 
 		if (new_offset < _current_offset) {
 			// wrap around in converting a VERY large long to an int
